requirements.txt
requests==2.25.1
urllib3==1.26.6
idna==2.10
chardet==4.0.0
certifi==2021.5.30
smmap==4.0.0
GitPython==3.1.18
gitdb==4.0.7
lxml==4.6.3